+++ src/main/java/com/giwankim/core/mvc/DispatcherServlet.java
@@ -1,7 +1,7 @@
 package com.giwankim.core.mvc;
 
-import com.giwankim.next.controller.UnauthorizedException;
-import com.giwankim.next.controller.UserNotFoundException;
+import com.giwankim.next.controller.user.UnauthorizedException;
+import com.giwankim.next.controller.user.UserNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
