+++ src/test/java/com/giwankim/next/controller/user/ProfileControllerTest.java
@@ -1,4 +1,4 @@
-package com.giwankim.next.controller;
+package com.giwankim.next.controller.user;
 
 import com.giwankim.next.dao.UserDao;
 import org.junit.jupiter.api.BeforeEach;
